// WYNKWeb/src/app/Models/MedicalBillMaster.model .ts
export class MedicalBill_Master {

  CMPID: number;
  RegistrationTranID: number;
  UIN: string;
  BillNo: string;
  TransactionId: number;
  GrossProductValue: number;
  TotalDiscountValue: number;
  TotalTaxValue: number;
  TotalCGSTTaxValue: number;
  TotalSGSTTaxValue: number;
  TotalBillValue: number;
  CreatedBy: number;
}