// WYNKWeb/src/app/Models/MedicalBillTran.model.ts
export class MedicalBillTran {

  MedicalBillID: number;
  DrugID: number;
  Quantity: number;
  MedicalPrescriptionID: number;
  MedicalBillParentID: number;
  IsMedicinePrescribed: boolean;
  ItemRate: number;
  ItemValue: number;
  UOM: string;
  DiscountPercentage: number;
  DiscountAmount: number;
  GSTPercentage: number;
  GSTTaxValue: number;
  CGSTPercentage: number;
  CGSTTaxValue: number;
  SGSTPercentage: number;
  SGSTTaxValue: number;
  CreatedBy: number;

}